Refactoring name of custom hook file, and using theme colouring spacing etc
